separate out into an individual factory
